Move utilities to server folder
