+++ packages/react-dom/src/server/escapeTextForBrowser.js
